% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_peak_expression.R
\name{by_peak_expression}
\alias{by_peak_expression}
\title{Compare Gene Expression by Peak Accessibility
Compares normalised gene expression between cells with and without accessibility at a linked peak. Includes total RNA UMI counts to check for the impact of library size. Returns a tibble by default or a ggplot object if \code{plot = TRUE}.}
\usage{
by_peak_expression(
  links_data,
  seurat_object = system.file("model_link_data.rds", pacckage = "linkinspectR"),
  plot = FALSE
)
}
\arguments{
\item{links_data}{A tibble generated by \code{get_link_data()}}

\item{seurat_object}{A Seurat object containing peak-to-gene links generated by \code{Signac::LinkPeaks()}}

\item{plot}{Logical; if \code{TRUE}, returns a ggplot object visualising expression and UMI distributions. Default is \code{FALSE}}
}
\value{
A tibble with one row per cell-peak-gene combination and the following columns:
\itemize{
\item \code{gene}: the selected linked gene
\item \code{peak}: the selected linked peak
\item \code{score}: the correlation coefficient - from LinkPeaks
\item \code{zscore}: the z-score of the correlation coefficient - from LinkPeaks
\item \code{pvalue}: the p-value associated with the z-score - from LinkPeaks
\item \code{cell_barcode}: the unique identifier for each cell - from Seurat Object
\item \code{gene_data}: the normalised RNAseq expression data - from Seurat Object
\item \code{gene_counts}: the raw RNAseq count data - from Seurat Object
\item \code{peak_data}: the normalised ATAC expression data - from Seurat Object
\item \code{peak_counts}: the raw ATAC count data - from Seurat Object
\item \code{read_in_peak}: Logical; \code{TRUE} if a read is observed in the peak for a given cell
\item \code{nCount_RNA}: Total RNA UMI counts per cell - from Seurat Object
}
}
\description{
Compare Gene Expression by Peak Accessibility
Compares normalised gene expression between cells with and without accessibility at a linked peak. Includes total RNA UMI counts to check for the impact of library size. Returns a tibble by default or a ggplot object if \code{plot = TRUE}.
}
