% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_link_data.R
\name{get_link_data}
\alias{get_link_data}
\title{Get Data Associated with Peak-to-Gene Links
Extracts data related to specified peak-to-gene links from a Seurat object, including gene expression and peak accessibility for each cell, and determines whether a read is present in the linked peak per cell.}
\usage{
get_link_data(
  seurat_object,
  Gene = NULL,
  Peak = NULL,
  Pvalue = NULL,
  expression.assay = "RNA",
  chromatin.assay = "ATAC"
)
}
\arguments{
\item{seurat_object}{A Seurat object containing peak-to-gene links generated by \code{Signac::LinkPeaks()}}

\item{Gene}{Character vector of gene name(s) of links to include}

\item{Peak}{Character vector of peak name(s) of links to include}

\item{Pvalue}{Numeric value specifying threshold p-value which links need to be lower than to be included}

\item{expression.assay}{Name of the assay containing gene expression information, default is RNA}

\item{chromatin.assay}{Name of the assay containing peak information, default is ATAC}
}
\value{
A tibble with one row per cell-peak-gene combination and the following columns:
\describe{
\item{gene}{: the selected linked gene}
\item{peak}{: the selected linked peak}
\item{score}{: the correlation coefficient - from LinkPeaks}
\item{zscore}{: the z-score of the correlation coefficient - from LinkPeaks}
\item{pvalue}{: the p-value associated with the z-score - from LinkPeaks}
\item{cell_barcode}{: the unique identifier for each cell - from Seurat Object}
\item{gene_data}{: the normalised RNAseq expression data - from Seurat Object}
\item{gene_counts}{: the raw RNAseq count data - from Seurat Object}
\item{peak_data}{: the normalised ATAC expression data - from Seurat Object}
\item{peak_counts}{: the raw ATAC count data - from Seurat Object}
\item{read_in_peak}{: Logical; \code{TRUE} if a read is observed in the peak for a given cell}
}
}
\description{
Get Data Associated with Peak-to-Gene Links
Extracts data related to specified peak-to-gene links from a Seurat object, including gene expression and peak accessibility for each cell, and determines whether a read is present in the linked peak per cell.
}
