% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkinspectR-package.R
\docType{package}
\name{linkinspectR-package}
\alias{linkinspectR-package}
\alias{linkinspectR}
\title{linkinspectR: A package for reviewing peak-to-gene links in multiomic single-cell datasets}
\description{
The linkinspectR package provides three main functions
}
\section{linkinspectR functions}{

get_link_data

by_peak_expression

by_peak_proportion
}

\author{
\strong{Maintainer}: Sarah Inglesfield \email{sarah.inglesfield@babraham.ac.uk} (\href{https://orcid.org/0000-0002-3511-9090}{ORCID})

}
